refactor, move workspace-provider to workspace-main-runtime
